+++ connector/src/test/java/it/eng/connector/integration/DataTransferTest.java
@@ -52,7 +52,7 @@ public class DataTransferTest extends BaseIntegrationTest {
 		TransferRequestMessage transferRequestMessage = TransferRequestMessage.Builder.newInstance()
 	    		.consumerPid(MockObjectUtil.CONSUMER_PID)
 	    		.agreementId("urn:uuid:AGREEMENT_ID_INITIALIZED")
-	    		.format(DataTransferFormat.HTTP_PULL.name())
+	    		.format(DataTransferFormat.HTTP_PULL.format())
 	    		.callbackAddress(MockObjectUtil.CALLBACK_ADDRESS)
 	    		.build();
 		
@@ -102,7 +102,7 @@ public class DataTransferTest extends BaseIntegrationTest {
 		TransferRequestMessage transferRequestMessage = TransferRequestMessage.Builder.newInstance()
 	    		.consumerPid(MockObjectUtil.CONSUMER_PID)
 	    		.agreementId("urn:uuid:AGREEMENT_ID") // this one should be present in init_data.json
-	    		.format(DataTransferFormat.HTTP_PULL.name())
+	    		.format(DataTransferFormat.HTTP_PULL.format())
 	    		.callbackAddress(MockObjectUtil.CALLBACK_ADDRESS)
 	    		.build();
 		
@@ -232,7 +232,7 @@ public class DataTransferTest extends BaseIntegrationTest {
 		TransferRequestMessage transferRequestMessage = TransferRequestMessage.Builder.newInstance()
 	    		.consumerPid(consumerPid)
 	    		.agreementId("urn:uuid:AGREEMENT_ID_COMPLETED_TRANSFER_TEST")
-	    		.format(DataTransferFormat.HTTP_PULL.name())
+	    		.format(DataTransferFormat.HTTP_PULL.format())
 	    		.callbackAddress(MockObjectUtil.CALLBACK_ADDRESS)
 	    		.build();
 		MvcResult mvcResult = mockMvc.perform(
@@ -346,7 +346,7 @@ public class DataTransferTest extends BaseIntegrationTest {
 		TransferRequestMessage transferRequestMessage = TransferRequestMessage.Builder.newInstance()
 	    		.consumerPid(consumerPid)
 	    		.agreementId("urn:uuid:AGREEMENT_ID_TERMINATE_TRANSFER_TEST")
-	    		.format(DataTransferFormat.HTTP_PULL.name())
+	    		.format(DataTransferFormat.HTTP_PULL.format())
 	    		.callbackAddress(MockObjectUtil.CALLBACK_ADDRESS)
 	    		.build();
 		MvcResult mvcResult = mockMvc.perform(